Altera tipo de Price para decimal

+++ src/Domain/Entities/Product.cs
@@ -1,8 +1,8 @@
 ﻿namespace Domain.Entities;
 
-public class Product(int id, string name, double price)
+public class Product(int id, string name, decimal price)
 {
     public int Id { get; } = id;
     public string Name { get; } = name;
-    public double Price { get; } = price;
+    public decimal Price { get; } = price;
 }
